Feat: New arrow key buttons and touch detection
I added two "E" and "Q" buttons to the HTML so I can use them for building more actions for our character. There's this new function called "touchDetectionAction" which detects if two objects (character and the space ship) are touching or not, if they do the E button will have an animation so the user knows it has to press it. There's also this function called gameLoop which keeps refreshing the touchDetectionAction function to check the detections